refactor: change package, sort classes, refactor files
