+++ Example/PrebidDemoJava/src/main/java/org/prebid/mobile/javademo/activities/MainActivity.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.prebid.mobile.app;
+package org.prebid.mobile.javademo.activities;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -26,9 +26,10 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.databinding.DataBindingUtil;
 import org.prebid.mobile.ExternalUserId;
 import org.prebid.mobile.PrebidMobile;
-import org.prebid.mobile.app.ads.AdType;
-import org.prebid.mobile.app.ads.AdTypesRepository;
-import org.prebid.mobile.app.databinding.ActivityMainBinding;
+import org.prebid.mobile.javademo.R;
+import org.prebid.mobile.javademo.ads.AdType;
+import org.prebid.mobile.javademo.ads.AdTypesRepository;
+import org.prebid.mobile.javademo.databinding.ActivityMainBinding;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -37,6 +38,13 @@ import java.util.Map;
 
 public class MainActivity extends AppCompatActivity {
 
+    /**
+     * This START fields help to set start values for selectors.
+     */
+    private static final String START_AD_SERVER = "Google Ad Manager (AWS)";
+    private static final String START_AD_TYPE = "";
+
+    private boolean isFirstInit = true;
     private String adType = "";
     private String adServer = "";
 
@@ -49,6 +57,7 @@ public class MainActivity extends AppCompatActivity {
         binding.btnShowAd.setOnClickListener((view) -> showAd());
         initPrebidExternalUserIds();
         initAdServerSpinner();
+        initStartServer();
     }
 
     public void showAd() {
@@ -69,38 +78,36 @@ public class MainActivity extends AppCompatActivity {
 
     private void initPrebidExternalUserIds() {
         ArrayList<ExternalUserId> externalUserIdArray = new ArrayList<>();
-        externalUserIdArray.add(new ExternalUserId(
-                "adserver.org",
-                "111111111111",
-                null,
-                new HashMap() {{
-                    put("rtiPartner", "TDID");
-                }}
-        ));
+        externalUserIdArray.add(new ExternalUserId("adserver.org", "111111111111", null, new HashMap() {{
+            put("rtiPartner", "TDID");
+        }}));
         externalUserIdArray.add(new ExternalUserId("netid.de", "999888777", null, null));
         externalUserIdArray.add(new ExternalUserId("criteo.com", "_fl7bV96WjZsbiUyQnJlQ3g4ckh5a1N", null, null));
         externalUserIdArray.add(new ExternalUserId("liveramp.com", "AjfowMv4ZHZQJFM8TpiUnYEyA81Vdgg", null, null));
-        externalUserIdArray.add(new ExternalUserId(
-                "sharedid.org",
-                "111111111111",
-                1,
-                new HashMap() {{
-                    put("third", "01ERJWE5FS4RAZKG6SKQ3ZYSKV");
-                }}
-        ));
+        externalUserIdArray.add(new ExternalUserId("sharedid.org", "111111111111", 1, new HashMap() {{
+            put("third", "01ERJWE5FS4RAZKG6SKQ3ZYSKV");
+        }}));
         PrebidMobile.setExternalUserIds(externalUserIdArray);
     }
 
     private void initAdServerSpinner() {
         Map<String, List<AdType>> repository = AdTypesRepository.get();
         ArrayList<String> primaryAdServers = new ArrayList<>(repository.keySet());
-        ArrayAdapter<String> adapter = new ArrayAdapter<>(this, android.R.layout.simple_spinner_dropdown_item, primaryAdServers);
+        ArrayAdapter<String> adapter = new ArrayAdapter<>(this,
+                android.R.layout.simple_spinner_dropdown_item,
+                primaryAdServers
+        );
 
         Spinner spinner = binding.spinnerAdServer;
         spinner.setAdapter(adapter);
         spinner.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener() {
             @Override
-            public void onItemSelected(AdapterView<?> parent, View view, int position, long id) {
+            public void onItemSelected(
+                    AdapterView<?> parent,
+                    View view,
+                    int position,
+                    long id
+            ) {
                 adServer = primaryAdServers.get(position);
                 List<AdType> adTypes = repository.get(adServer);
                 ArrayList<String> stringTypes = new ArrayList<>(5);
@@ -123,14 +130,56 @@ public class MainActivity extends AppCompatActivity {
         spinner.setAdapter(adapter);
         spinner.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener() {
             @Override
-            public void onItemSelected(AdapterView<?> parent, View view, int position, long id) {
+            public void onItemSelected(
+                    AdapterView<?> parent,
+                    View view,
+                    int position,
+                    long id
+            ) {
                 adType = list.get(position);
+                initStartAdType();
             }
 
             @Override
             public void onNothingSelected(AdapterView<?> parent) {
             }
         });
+    }
+
+    private void initStartServer() {
+        ArrayList<String> keySet = new ArrayList<>(AdTypesRepository.get().keySet());
+        int indexOfServer = -1;
+        for (int i = 0; i < keySet.size(); i++) {
+            String key = keySet.get(i);
+            if (key.equalsIgnoreCase(START_AD_SERVER)) {
+                indexOfServer = i;
+            }
+        }
+
+        if (indexOfServer >= 0) {
+            binding.spinnerAdServer.setSelection(indexOfServer, false);
+        } else {
+            isFirstInit = false;
+        }
+    }
+
+    private void initStartAdType() {
+        if (isFirstInit) {
+            isFirstInit = false;
+
+            try {
+                List<AdType> adTypes = AdTypesRepository.get().get(START_AD_TYPE);
+                if (adTypes != null) {
+                    for (int i = 0; i < adTypes.size(); i++) {
+                        AdType adType = adTypes.get(i);
+                        if (adType.getName().equalsIgnoreCase(START_AD_TYPE)) {
+                            binding.spinnerAdType.setSelection(i, false);
+                        }
+                    }
+                }
+            } catch (Exception ignored) {
+            }
+        }
     }
 
 }
