+++ Example/PrebidDemoJava/src/main/java/org/prebid/mobile/javademo/ads/rubicon/RubiconGamInterstitial.java
@@ -1,4 +1,4 @@
-package org.prebid.mobile.app.ads.rubicon;
+package org.prebid.mobile.javademo.ads.rubicon;
 
 import android.app.Activity;
 import android.os.Build;
@@ -8,15 +8,12 @@ import com.google.android.gms.ads.LoadAdError;
 import com.google.android.gms.ads.admanager.AdManagerAdRequest;
 import com.google.android.gms.ads.admanager.AdManagerInterstitialAd;
 import com.google.android.gms.ads.admanager.AdManagerInterstitialAdLoadCallback;
-import org.prebid.mobile.Signals;
-import org.prebid.mobile.VideoBaseAdUnit;
-import org.prebid.mobile.VideoInterstitialAdUnit;
+import org.prebid.mobile.AdUnit;
+import org.prebid.mobile.InterstitialAdUnit;
 
-import java.util.Collections;
+public class RubiconGamInterstitial {
 
-public class RubiconGamVideoInterstitial {
-
-    private static VideoInterstitialAdUnit adUnit;
+    public static AdUnit adUnit;
 
     public static void create(
             Activity activity,
@@ -24,20 +21,12 @@ public class RubiconGamVideoInterstitial {
             String configId,
             int autoRefreshTime
     ) {
-        VideoBaseAdUnit.Parameters parameters = new VideoBaseAdUnit.Parameters();
-        parameters.setMimes(Collections.singletonList("video/mp4"));
-        parameters.setProtocols(Collections.singletonList(Signals.Protocols.VAST_2_0));
-        parameters.setPlaybackMethod(Collections.singletonList(Signals.PlaybackMethod.AutoPlaySoundOff));
-
-        adUnit = new VideoInterstitialAdUnit(configId);
-        adUnit.setParameters(parameters);
+        adUnit = new InterstitialAdUnit(configId);
         adUnit.setAutoRefreshInterval(autoRefreshTime);
 
         final AdManagerAdRequest.Builder builder = new AdManagerAdRequest.Builder();
-
         adUnit.fetchDemand(builder, resultCode -> {
             AdManagerAdRequest request = builder.build();
-
             AdManagerInterstitialAd.load(activity, adUnitId, request, createListener(activity));
         });
     }
@@ -52,25 +41,25 @@ public class RubiconGamVideoInterstitial {
     private static AdManagerInterstitialAdLoadCallback createListener(Activity activity) {
         return new AdManagerInterstitialAdLoadCallback() {
             @Override
-            public void onAdLoaded(@NonNull AdManagerInterstitialAd adManagerInterstitialAd) {
-                adManagerInterstitialAd.show(activity);
+            public void onAdLoaded(@NonNull AdManagerInterstitialAd interstitialManager) {
+                interstitialManager.show(activity);
             }
 
             @Override
             public void onAdFailedToLoad(@NonNull LoadAdError loadAdError) {
-                showDialog(activity, loadAdError.getMessage());
+                createDialog(activity, loadAdError.getMessage());
             }
         };
     }
 
-    private static void showDialog(Activity activity, String message) {
-        AlertDialog.Builder builder;
+    private static void createDialog(Activity activity, String message) {
+        AlertDialog.Builder builder1;
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-            builder = new AlertDialog.Builder(activity, android.R.style.Theme_Material_Dialog_Alert);
+            builder1 = new AlertDialog.Builder(activity, android.R.style.Theme_Material_Dialog_Alert);
         } else {
-            builder = new AlertDialog.Builder(activity);
+            builder1 = new AlertDialog.Builder(activity);
         }
-        builder.setTitle("Failed to load AdManager interstitial ad")
+        builder1.setTitle("Failed to load AdManager interstitial ad")
                 .setMessage("Error: " + message)
                 .setIcon(android.R.drawable.ic_dialog_alert)
                 .show();
