+++ Example/PrebidDemoJava/src/main/java/org/prebid/mobile/javademo/ads/rubicon/RubiconGamVideoInterstitial.java
@@ -1,4 +1,4 @@
-package org.prebid.mobile.app.ads.rubicon;
+package org.prebid.mobile.javademo.ads.rubicon;
 
 import android.app.Activity;
 import android.os.Build;
@@ -8,12 +8,15 @@ import com.google.android.gms.ads.LoadAdError;
 import com.google.android.gms.ads.admanager.AdManagerAdRequest;
 import com.google.android.gms.ads.admanager.AdManagerInterstitialAd;
 import com.google.android.gms.ads.admanager.AdManagerInterstitialAdLoadCallback;
-import org.prebid.mobile.AdUnit;
-import org.prebid.mobile.InterstitialAdUnit;
+import org.prebid.mobile.Signals;
+import org.prebid.mobile.VideoBaseAdUnit;
+import org.prebid.mobile.VideoInterstitialAdUnit;
 
-public class RubiconGamInterstitial {
+import java.util.Collections;
 
-    public static AdUnit adUnit;
+public class RubiconGamVideoInterstitial {
+
+    private static VideoInterstitialAdUnit adUnit;
 
     public static void create(
             Activity activity,
@@ -21,12 +24,20 @@ public class RubiconGamInterstitial {
             String configId,
             int autoRefreshTime
     ) {
-        adUnit = new InterstitialAdUnit(configId);
+        VideoBaseAdUnit.Parameters parameters = new VideoBaseAdUnit.Parameters();
+        parameters.setMimes(Collections.singletonList("video/mp4"));
+        parameters.setProtocols(Collections.singletonList(Signals.Protocols.VAST_2_0));
+        parameters.setPlaybackMethod(Collections.singletonList(Signals.PlaybackMethod.AutoPlaySoundOff));
+
+        adUnit = new VideoInterstitialAdUnit(configId);
+        adUnit.setParameters(parameters);
         adUnit.setAutoRefreshInterval(autoRefreshTime);
 
         final AdManagerAdRequest.Builder builder = new AdManagerAdRequest.Builder();
+
         adUnit.fetchDemand(builder, resultCode -> {
             AdManagerAdRequest request = builder.build();
+
             AdManagerInterstitialAd.load(activity, adUnitId, request, createListener(activity));
         });
     }
@@ -41,25 +52,25 @@ public class RubiconGamInterstitial {
     private static AdManagerInterstitialAdLoadCallback createListener(Activity activity) {
         return new AdManagerInterstitialAdLoadCallback() {
             @Override
-            public void onAdLoaded(@NonNull AdManagerInterstitialAd interstitialManager) {
-                interstitialManager.show(activity);
+            public void onAdLoaded(@NonNull AdManagerInterstitialAd adManagerInterstitialAd) {
+                adManagerInterstitialAd.show(activity);
             }
 
             @Override
             public void onAdFailedToLoad(@NonNull LoadAdError loadAdError) {
-                createDialog(activity, loadAdError.getMessage());
+                showDialog(activity, loadAdError.getMessage());
             }
         };
     }
 
-    private static void createDialog(Activity activity, String message) {
-        AlertDialog.Builder builder1;
+    private static void showDialog(Activity activity, String message) {
+        AlertDialog.Builder builder;
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-            builder1 = new AlertDialog.Builder(activity, android.R.style.Theme_Material_Dialog_Alert);
+            builder = new AlertDialog.Builder(activity, android.R.style.Theme_Material_Dialog_Alert);
         } else {
-            builder1 = new AlertDialog.Builder(activity);
+            builder = new AlertDialog.Builder(activity);
         }
-        builder1.setTitle("Failed to load AdManager interstitial ad")
+        builder.setTitle("Failed to load AdManager interstitial ad")
                 .setMessage("Error: " + message)
                 .setIcon(android.R.drawable.ic_dialog_alert)
                 .show();
