+++ Example/PrebidDemoJava/src/main/java/org/prebid/mobile/javademo/utils/Util.java
@@ -1,10 +1,11 @@
-package org.prebid.mobile.app;
+package org.prebid.mobile.javademo.utils;
 
 import org.prebid.mobile.AdUnit;
 import org.prebid.mobile.PrebidMobile;
 import org.prebid.mobile.TargetingParams;
 
 public class Util {
+
     public static void enableAdditionalFunctionality(AdUnit adUnit) {
         enableCOPPA();
         addFirstPartyData(adUnit);
@@ -18,33 +19,29 @@ public class Util {
     }
 
     private static void addFirstPartyData(AdUnit adUnit) {
-        //Access Control List
+        /* Access Control List */
         TargetingParams.addBidderToAccessControlList(TargetingParams.BIDDER_NAME_RUBICON_PROJECT);
-
-        //PBAdSlot(should be set together with Access Control List)
+        /* PBAdSlot(should be set together with Access Control List) */
         adUnit.setPbAdSlot("/1111111/homepage/med-rect-2");
 
-        //global user data
+        /* Global user data */
         TargetingParams.addUserData("globalUserDataKey1", "globalUserDataValue1");
-
-        //global context data
+        /* Global context data */
         TargetingParams.addContextData("globalContextDataKey1", "globalContextDataValue1");
-
-        //adunit context data
+        /* Ad unit context data */
         adUnit.addContextData("adunitContextDataKey1", "adunitContextDataValue1");
 
-        //global context keywords
+        /* Context keywords */
         TargetingParams.addContextKeyword("globalContextKeywordValue1");
         TargetingParams.addContextKeyword("globalContextKeywordValue2");
 
-        //global user keywords
+        /* Global user keywords */
         TargetingParams.addUserKeyword("globalUserKeywordValue1");
         TargetingParams.addUserKeyword("globalUserKeywordValue2");
 
-        //adunit context keywords
+        /* Ad unit context keywords */
         adUnit.addContextKeyword("adunitContextKeywordValue1");
         adUnit.addContextKeyword("adunitContextKeywordValue2");
-
     }
 
     private static void setStoredResponse() {
@@ -58,4 +55,5 @@ public class Util {
     private static void enablePbsDebug() {
         PrebidMobile.setPbsDebug(true);
     }
+
 }
