+++ Example/PrebidDemoJava/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
-        tools:context="org.prebid.mobile.app.MainActivity">
+        tools:context="org.prebid.mobile.javademo.activities.MainActivity">
 
         <LinearLayout
             android:layout_width="match_parent"
